Add support to function calls on `colon` rule
