+++ CHANGELOG.md
@@ -11,7 +11,7 @@
   [#1834](https://github.com/realm/SwiftLint/issues/1834)
 
 * Improve `colon` rule to catch violations in dictionary types
-  (e.g. `[String: Int]`) and when using `Any`.  
+  (e.g. `[String: Int]`), when using `Any` and on function calls.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#1074](https://github.com/realm/SwiftLint/issues/1811)
