+++ Source/SwiftLintFramework/Rules/ColonRule.swift
@@ -12,9 +12,10 @@ import SourceKittenFramework
 private enum ColonKind {
     case type
     case dictionary
+    case functionCall
 }
 
-public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
+public struct ColonRule: CorrectableRule, ConfigurationProviderRule {
 
     public var configuration = ColonConfiguration()
 
@@ -53,7 +54,11 @@ public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
             "switch foo {\n" +
             "case .bar:\n" +
             "    _ = something()\n" +
-            "}\n"
+            "}\n",
+            "object.method(x: 5, y: \"string\")\n",
+            "object.method(x: 5, y:\n" +
+            "              \"string\")",
+            "object.method(5, y: \"string\")\n"
         ],
         triggeringExamples: [
             "let ↓abc:Void\n",
@@ -96,7 +101,10 @@ public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
             "class ↓Foo : Bar {}\n",
             "class ↓Foo:Bar {}\n",
             "class Foo<↓T:Equatable> {}\n",
-            "class Foo<↓T : Equatable> {}\n"
+            "class Foo<↓T : Equatable> {}\n",
+            "object.method(x: 5, y↓ : \"string\")\n",
+            "object.method(x↓:5, y: \"string\")\n",
+            "object.method(x↓:  5, y: \"string\")\n"
         ],
         corrections: [
             "let ↓abc:Void\n": "let abc: Void\n",
@@ -139,7 +147,10 @@ public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
             "class ↓Foo : Bar {}\n": "class Foo: Bar {}\n",
             "class ↓Foo:Bar {}\n": "class Foo: Bar {}\n",
             "class Foo<↓T:Equatable> {}\n": "class Foo<T: Equatable> {}\n",
-            "class Foo<↓T : Equatable> {}\n": "class Foo<T: Equatable> {}\n"
+            "class Foo<↓T : Equatable> {}\n": "class Foo<T: Equatable> {}\n",
+            "object.method(x: 5, y↓ : \"string\")\n": "object.method(x: 5, y: \"string\")\n",
+            "object.method(x↓:5, y: \"string\")\n": "object.method(x: 5, y: \"string\")\n",
+            "object.method(x↓:  5, y: \"string\")\n": "object.method(x: 5, y: \"string\")\n"
         ]
     )
 
@@ -178,7 +189,7 @@ public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
                                                                       options: [],
                                                                       range: range,
                                                                       withTemplate: "$1: $2")
-            case .dictionary:
+            case .dictionary, .functionCall:
                 contents = contents.bridge().replacingCharacters(in: range, with: ": ")
             }
 
@@ -201,94 +212,31 @@ public struct ColonRule: ASTRule, CorrectableRule, ConfigurationProviderRule {
             guard let range = contents.byteRangeToNSRange(start: $0.location, length: $0.length) else {
                 return nil
             }
-            return (range: range, kind: ColonKind.dictionary)
+            return (range: range, kind: .dictionary)
+        }
+        let functionViolations: [RangeWithKind] = functionCallColonViolationRanges(in: file,
+                                                                                   dictionary: dictionary).flatMap {
+            guard let range = contents.byteRangeToNSRange(start: $0.location, length: $0.length) else {
+                return nil
+            }
+            return (range: range, kind: .functionCall)
         }
 
-        return (violations + dictViolations).sorted { $0.range.location < $1.range.location }
-    }
-}
-
-// MARK: Type Colon Rule
-
-private extension ColonRule {
-
-    var pattern: String {
-        // If flexible_right_spacing is true, match only 0 whitespaces.
-        // If flexible_right_spacing is false or omitted, match 0 or 2+ whitespaces.
-        let spacingRegex = configuration.flexibleRightSpacing ? "(?:\\s{0})" : "(?:\\s{0}|\\s{2,})"
-
-        return "(\\w)" +       // Capture an identifier
-               "(?:" +         // start group
-               "\\s+" +        // followed by whitespace
-               ":" +           // to the left of a colon
-               "\\s*" +        // followed by any amount of whitespace.
-               "|" +           // or
-               ":" +           // immediately followed by a colon
-               spacingRegex +  // followed by right spacing regex
-               ")" +           // end group
-               "(" +           // Capture a type identifier
-               "[\\[|\\(]*" +  // which may begin with a series of nested parenthesis or brackets
-               "\\S)"          // lazily to the first non-whitespace character.
-    }
-
-    func typeColonViolationRanges(in file: File, matching pattern: String) -> [NSRange] {
-        let nsstring = file.contents.bridge()
-        let commentAndStringKindsSet = Set(SyntaxKind.commentAndStringKinds())
-        return file.rangesAndTokens(matching: pattern).filter { _, syntaxTokens in
-            let syntaxKinds = syntaxTokens.flatMap { SyntaxKind(rawValue: $0.type) }
-
-            guard syntaxKinds.count == 2 else {
-                return false
-            }
-
-            let validKinds: Bool
-            switch (syntaxKinds[0], syntaxKinds[1]) {
-            case (.identifier, .typeidentifier),
-                 (.typeidentifier, .typeidentifier):
-                validKinds = true
-            case (.identifier, .keyword),
-                 (.typeidentifier, .keyword):
-                validKinds = file.isTypeLike(token: syntaxTokens[1])
-            case (.keyword, .typeidentifier):
-                validKinds = file.isTypeLike(token: syntaxTokens[0])
-            default:
-                validKinds = false
-            }
-
-            guard validKinds else {
-                return false
-            }
-
-            return Set(syntaxKinds).isDisjoint(with: commentAndStringKindsSet)
-        }.flatMap { range, syntaxTokens in
-            let identifierRange = nsstring
-                .byteRangeToNSRange(start: syntaxTokens[0].offset, length: 0)
-            return identifierRange.map { NSUnionRange($0, range) }
+        return (violations + dictViolations + functionViolations).sorted {
+            $0.range.location < $1.range.location
         }
     }
 }
 
-private extension File {
-    func isTypeLike(token: SyntaxToken) -> Bool {
-        let nsstring = contents.bridge()
-        guard let text = nsstring.substringWithByteRange(start: token.offset, length: token.length),
-            let firstLetter = text.unicodeScalars.first else {
-                return false
-        }
+extension ColonRule: ASTRule {
 
-        return CharacterSet.uppercaseLetters.contains(firstLetter)
-    }
-}
-
-// MARK: Dictionary Colon Rule
-
-extension ColonRule {
-
-    /// Only returns dictionary colon violations
+    /// Only returns dictionary and function calls colon violations
     public func validate(file: File, kind: SwiftExpressionKind,
                          dictionary: [String: SourceKitRepresentable]) -> [StyleViolation] {
 
-        let ranges = dictionaryColonViolationRanges(in: file, kind: kind, dictionary: dictionary)
+        let ranges = dictionaryColonViolationRanges(in: file, kind: kind, dictionary: dictionary) +
+            functionCallColonViolationRanges(in: file, kind: kind, dictionary: dictionary)
+
         return ranges.map {
             StyleViolation(ruleDescription: type(of: self).description,
                            severity: configuration.severityConfiguration.severity,
@@ -296,70 +244,4 @@ extension ColonRule {
         }
     }
 
-    fileprivate func dictionaryColonViolationRanges(in file: File,
-                                                    dictionary: [String: SourceKitRepresentable]) -> [NSRange] {
-        guard configuration.applyToDictionaries else {
-            return []
-        }
-
-        return dictionary.substructure.flatMap { subDict -> [NSRange] in
-            guard let kindString = subDict.kind,
-                let kind = KindType(rawValue: kindString) else {
-                    return []
-            }
-            return dictionaryColonViolationRanges(in: file, dictionary: subDict) +
-                dictionaryColonViolationRanges(in: file, kind: kind, dictionary: subDict)
-        }
-    }
-
-    private func dictionaryColonViolationRanges(in file: File, kind: SwiftExpressionKind,
-                                                dictionary: [String: SourceKitRepresentable]) -> [NSRange] {
-        guard kind == .dictionary,
-            let ranges = colonRanges(dictionary: dictionary) else {
-                return []
-        }
-
-        let contents = file.contents.bridge()
-        return ranges.filter {
-            guard let colon = contents.substringWithByteRange(start: $0.location,
-                                                              length: $0.length) else {
-                return false
-            }
-
-            if configuration.flexibleRightSpacing {
-                let isCorrect = colon.hasPrefix(": ") || colon.hasPrefix(":\n")
-                return !isCorrect
-            }
-
-            return colon != ": " && !colon.hasPrefix(":\n")
-        }
-    }
-
-    private func colonRanges(dictionary: [String: SourceKitRepresentable]) -> [NSRange]? {
-        let elements = dictionary.elements
-        guard elements.count % 2 == 0 else {
-                return nil
-        }
-
-        let expectedKind = "source.lang.swift.structure.elem.expr"
-        let ranges: [NSRange] = elements.flatMap { subDict in
-            guard subDict.kind == expectedKind,
-                let offset = subDict.offset,
-                let length = subDict.length else {
-                    return nil
-            }
-
-            return NSRange(location: offset, length: length)
-        }
-
-        let even = ranges.enumerated().flatMap { $0 % 2 == 0 ? $1 : nil }
-        let odd = ranges.enumerated().flatMap { $0 % 2 != 0 ? $1 : nil }
-
-        return zip(even, odd).map { evenRange, oddRange -> NSRange in
-            let location = NSMaxRange(evenRange)
-            let length = oddRange.location - location
-
-            return NSRange(location: location, length: length)
-        }
-    }
 }
